QS-935: Refactor enterprise users controllers

+++ src/routing.php
@@ -177,27 +177,27 @@ $app->post('/groups/validate', 'admin.groups.controller:validateAction');
 $app->get('/invitations', 'admin.invitations.controller:indexAction');
 
 /** EnterpriseUser routes */
-$app->get('/enterpriseUsers/{id}', 'enterpriseUsers.controller:getAction');
-$app->post('/enterpriseUsers', 'enterpriseUsers.controller:postAction');
-$app->put('/enterpriseUsers/{id}', 'enterpriseUsers.controller:putAction');
-$app->delete('/enterpriseUsers/{id}', 'enterpriseUsers.controller:deleteAction');
-$app->post('/enterpriseUsers/{id}', 'enterpriseUsers.controller:validateAction');
+$app->get('/enterpriseUsers/{id}', 'admin.enterpriseUsers.controller:getAction');
+$app->post('/enterpriseUsers', 'admin.enterpriseUsers.controller:postAction');
+$app->put('/enterpriseUsers/{id}', 'admin.enterpriseUsers.controller:putAction');
+$app->delete('/enterpriseUsers/{id}', 'admin.enterpriseUsers.controller:deleteAction');
+$app->post('/enterpriseUsers/{id}', 'admin.enterpriseUsers.controller:validateAction');
 
 /** EnterpriseUser Group routes */
-$app->get('/enterpriseUsers/{enterpriseUserId}/groups', 'enterpriseUsers.groups.controller:getAllAction');
-$app->get('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'enterpriseUsers.groups.controller:getAction');
-$app->post('/enterpriseUsers/{enterpriseUserId}/groups', 'enterpriseUsers.groups.controller:postAction');
-$app->put('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'enterpriseUsers.groups.controller:putAction');
-$app->delete('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'enterpriseUsers.groups.controller:deleteAction');
-$app->post('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'enterpriseUsers.groups.controller:validateAction');
-$app->get('/enterpriseUsers/groups/{id}/communities', 'enterpriseUsers.communities.controller:getByGroupAction');
+$app->get('/enterpriseUsers/{enterpriseUserId}/groups', 'admin.enterpriseUsers.groups.controller:getAllAction');
+$app->get('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'admin.enterpriseUsers.groups.controller:getAction');
+$app->post('/enterpriseUsers/{enterpriseUserId}/groups', 'admin.enterpriseUsers.groups.controller:postAction');
+$app->put('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'admin.enterpriseUsers.groups.controller:putAction');
+$app->delete('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'admin.enterpriseUsers.groups.controller:deleteAction');
+$app->post('/enterpriseUsers/{enterpriseUserId}/groups/{id}', 'admin.enterpriseUsers.groups.controller:validateAction');
+$app->get('/enterpriseUsers/groups/{id}/communities', 'admin.enterpriseUsers.communities.controller:getByGroupAction');
 
 /** EnterpriseUser Invitation routes */
-$app->post('/enterpriseUsers/{enterpriseUserId}/invitations', 'enterpriseUsers.invitations.controller:postAction');
-$app->delete('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'enterpriseUsers.invitations.controller:deleteAction');
-$app->get('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'enterpriseUsers.invitations.controller:getAction');
-$app->put('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'enterpriseUsers.invitations.controller:putAction');
-$app->post('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'enterpriseUsers.invitations.controller:validateAction');
+$app->post('/enterpriseUsers/{enterpriseUserId}/invitations', 'admin.enterpriseUsers.invitations.controller:postAction');
+$app->delete('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'admin.enterpriseUsers.invitations.controller:deleteAction');
+$app->get('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'admin.enterpriseUsers.invitations.controller:getAction');
+$app->put('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'admin.enterpriseUsers.invitations.controller:putAction');
+$app->post('/enterpriseUsers/{enterpriseUserId}/invitations/{id}', 'admin.enterpriseUsers.invitations.controller:validateAction');
 
 $controllers
     ->assert('id', '\d+')
